Ported App Service example to Python

+++ app-service/service.destroy.sh
@@ -16,6 +16,8 @@ fi
 export RESOURCE_GROUP=${PREFIX}resourcegroup
 export STORAGE_ACCOUNT=${PREFIX}storageaccount
 export CONTAINER_NAME=${PREFIX}container
+export FILES_SHARE_NAME=${PREFIX}share
+export CONSUMPTION_PLAN=${PREFIX}consumptionplan
 export APP=${PREFIX}app
 export AD_NAME=${PREFIX}ad
 
@@ -24,16 +26,46 @@ REAL=$(realpath "$SELF")
 BASE=$(dirname "$REAL")
 
 # az login
-# Destroy Authentication / Authorization of the function app
-az ad app delete --id $(az ad app list --query "[?displayName == '$AD_NAME'].appId" --all --output tsv)
 # Destroy the function app
-az functionapp delete --name $APP --resource-group $RESOURCE_GROUP
+APP_RESULT=$(az functionapp list --resource-group $RESOURCE_GROUP --query "contains([].name,'$APP')")
+if [ "$APP_RESULT" = "true" ]; then
+    az functionapp delete --name $APP --resource-group $RESOURCE_GROUP
+fi
+# Destroy the consumption plan
+CONSUMPTION_PLAN_RESULT=$(az functionapp plan list --query "contains([].name, '$CONSUMPTION_PLAN')")
+if [ "$CONSUMPTION_PLAN_RESULT" = "true" ]; then
+    az appservice plan delete --resource-group $RESOURCE_GROUP --name $CONSUMPTION_PLAN -y
+fi
+# Destroy Authentication / Authorization of the function app
+AAD_ID=$(az ad app list --query "[?displayName == '$AD_NAME'].appId" --all --output tsv)
+if [ "$AAD_ID" != "" ]; then
+    az ad app delete --id $(az ad app list --query "[?displayName == '$AD_NAME'].appId" --all --output tsv)
+fi
+# Destroy the files volume
+STORAGE_ACCOUNT_CONNECTION_STRING=$(az storage account show-connection-string --resource-group $RESOURCE_GROUP --name $STORAGE_ACCOUNT -o tsv)
+FILES_SHARE_RESULT=$(az storage share list --account-name $STORAGE_ACCOUNT --connection-string $STORAGE_ACCOUNT_CONNECTION_STRING --query "contains([].name, '$FILES_SHARE_NAME')")
+if [ "$FILES_SHARE_RESULT" = "true" ]; then
+    az storage share delete --name $FILES_SHARE_NAME --connection-string $STORAGE_ACCOUNT_CONNECTION_STRING
+fi
 # Destroy the static content container
-az storage container delete --name $CONTAINER_NAME --account-name $STORAGE_ACCOUNT
+CONTAINER_RESULT=$(az storage container list --account-name $STORAGE_ACCOUNT --connection-string $STORAGE_ACCOUNT_CONNECTION_STRING --query "contains([].name, '$CONTAINER_NAME')")
+if [ "$CONTAINER_RESULT" = "true" ]; then
+    az storage container delete --name $CONTAINER_NAME --account-name $STORAGE_ACCOUNT --connection-string $STORAGE_ACCOUNT_CONNECTION_STRING
+fi
 # Destroy the storage account
-az storage account delete --name $STORAGE_ACCOUNT --resource-group $RESOURCE_GROUP -y
+STORAGE_ACCOUNT_RESULT=$(az storage account check-name --name $STORAGE_ACCOUNT --query nameAvailable)
+if [ "$STORAGE_ACCOUNT_RESULT" = "false" ]; then
+    az storage account delete --name $STORAGE_ACCOUNT --resource-group $RESOURCE_GROUP -y
+fi
 # Destroy the resource group
-az group delete --name $RESOURCE_GROUP -y
+RESOURCE_GROUP_RESULT=$(az group exists --name $RESOURCE_GROUP)
+if [ "$RESOURCE_GROUP_RESULT" = "true" ]; then
+    az group delete --name $RESOURCE_GROUP -y
+fi
 # Clean-up
-rm -rf $BASE/.azurefunctions || true
-rm $BASE/proxies.json $BASE/function.zip $BASE/swagger-client.js || true
+rm $BASE/app/proxies.json 2> /dev/null || true
+rm $BASE/wwwroot/swagger.json 2> /dev/null || true
+rm $BASE/wwwroot/swagger-client.js 2> /dev/null || true
+rm $BASE/date.txt 2> /dev/null || true
+rm -rf $BASE/app/.python_packages 2> /dev/null || true
+rm -rf $BASE/app/azure-functions-core-tools 2> /dev/null || true
